+++ tests/org/elixir_lang/parser_definition/Issue4068TestCase.kt
@@ -122,7 +122,7 @@ class Issue4068TestCase : ParsingTestCase() {
         assertParsedAndQuotedCorrectlyBeforeOrRaise(QuotingDialect.V1_18, QuotingDialect.V1_19, "Elixir.MatchError", false)
     fun testQuotedRemoteCallNameInvalidBracedEscape() =
         assertParsedAndQuotedCorrectlyBeforeOrRaise(QuotingDialect.V1_18, QuotingDialect.V1_19, "Elixir.MatchError", false)
-    fun testInvalidHexadecimalEscapeString() = assertParsedWithErrors(false)
+    fun testInvalidHexadecimalEscapeString() = assertParsedAndQuotedAroundError(false)
     fun testQuotedRemoteCallNameInvalidCodePoint() =
         assertParsedAndQuotedCorrectlyBeforeOrRaise(QuotingDialect.V1_18, QuotingDialect.V1_19, "Elixir.MatchError", false)
     fun testQuotedRemoteCallNameSurrogate() =